examples: linux: Factor out get_rpmsg_chrdev_fd into common file
