+++ examples/linux/common/common.h
@@ -12,5 +12,6 @@ char *get_rpmsg_ept_dev_name(const char *rpmsg_char_name,
                              const char *ept_name,
                              char *ept_dev_name);
 int bind_rpmsg_chrdev(const char *rpmsg_dev_name);
+int get_rpmsg_chrdev_fd(const char *rpmsg_dev_name, char *rpmsg_ctrl_name);
 
 #endif /* __COMMON__H__ */
